Moving classes and interface into their own packages.
