+++ src/DataProcessing/DataProcessing.java
@@ -1,0 +1,14 @@
+package src.DataProcessing;
+
+import src.encdec.IEncDecAlgorithm;
+
+abstract class DataProcessing {
+
+    protected IEncDecAlgorithm algorithm;
+
+    public DataProcessing(IEncDecAlgorithm algorithm) {
+        this.algorithm = algorithm;
+    }
+
+    protected abstract String processData(String data, int key);
+}
