+++ src/EncDecAlgorithms/IEncDecAlgorithm.java
@@ -1,0 +1,6 @@
+package src.EncDecAlgorithms;
+
+interface IEncDecAlgorithm {
+    String encrypt(String msg, int key);
+    String decrypt(String msg, int key);
+}
